Update AsteroidGameView.java
When the user says the correct word that appears on the asteroid traveling towards the rocket (player), the asteroid is destroyed and a new asteroid with a new word is created. It's really basic on the front-end, but I also added some code for cases in which there will be more asteroids located on the screen, though this is not shown to the player as of now.