Asteroid mini game (#15)
* Update AsteroidGameView.java

Added new explosions when a word is pronounced correctly and the associated asteroid is destroyed.

* Update AsteroidGameModel.java

- Eliminated unnecessary constants

* Update AsteroidGameController.java, AsteroidGameModel.java

- Added more variables and methods to the asteroid game model to allow for multiple asteroids
- Can determine how many words each level will have
- Can determine how many levels are possible if there is a small vocab list

* Update AsteroidGameController.java

- Added necessary get() methods

* Update AsteroidGameView.java

Removed unnecessary import statements

* Update AsteroidGameModel.java, AsteroidModel.java, AsteroidGameView.java

Multiple asteroids with their respective vocabulary words are displayed on the screen. The number of asteroids does not increase as progress is made in the mini-game yet, but will be implemented.

* Update AsteroidGameView.java

Added some code to allow for the spoken word to be compared to the words on every asteroid instead of just the first asteroid. It's a work in progress because I'm still working out the bugs.

* Update AsteroidGameView.java

The player can now say any word that is present on any asteroid on the screen and if it is pronounced correctly, it will be destroyed.

* Update AsteroidGameView.java

- Fixed a bug with the losing screen. The app crashed when the player lost the asteroid mini-game, but now it shows the losing screen and allows the player to navigate back to the main menu.
- Added a winning screen that appears when the player wins the game.

* Update AsteroidGameView.java, MainMenuScreen.java, Remove StartAsteroidGameController.java, StartWordScrambleGameController.java, Add asteroid_mini_game.md

- Updated some comments in AsteroidGameView.java and MainMenuScreen.java
- Removed files that are no longer in use
- Added asteroid_mini_game.md which includes documentation for the asteroid mini-game for future developers